package provider: add support for `source` parameter
Allow using the `source` parameter as an URI to determine from where to
install the package.

If `source` is not set, then use `pkg install` to install the package
using the default repository logic.

A URN of the form: urn:freebsd:repo:<tag> will force `pkg install` to
install from the given repository using the '-r' flag. URLs are passed
as-is to `pkg add` to install directly.