Purge default repos if specified (new default)
It's sensible that we want to do this as we don't really want to have
repos defined in both /usr/local/etc/ and /etc/

This changes the default and therefore the README - this behaviour is
what people want. We could not change the default, but I believe this
causes more confusion than anything else... so let's just not let people
do silly things to their *BSD boxes.