Fixes facts for when pkg was not installed by pkg.
Pkg may be installed via make or pkg_add. There is also a bug I am
trying to run down where pkg bootstraps fine but still isn't present in
`pkg info`.

This change fixes puppet facts for FBSD 10 or 11 in any of the
aforementioned cases. I did not touch FBSD 8 or 9, due to the warning
present in the pkg(8) description of the -N option.