Update pkgng provider from upstream 4.2.2
The provider has been merged to Puppet core, and some updates were made
to get it there.  This brings those changes back into the module for the
purpose of validation.  Once FreeBSD has a port for Puppet 4 that
includes the provider, we will remove the provider from this
module entirely.  As such, this update ensures that the provider in the
module today is the same as the provider that will ships with core.