Drop legacy code
WITH_PKGNG=yes was required for FreeBSD versions earlier than 10.x, and
this module supports FreeBSD 10 or 11, see:
https://www.freebsd.org/doc/handbook/pkgng-intro.html#pkgng-initial-setup

Likewisely, pkg2ng was needed for switching to pkg, which does not look
like a responsibility for a configuration management system like puppet
but for an orchestration system like MCollective.  It was unused by the
way.