Do not erase /etc/pkg/FreeBSD.conf
Erasing this file was unexpected, and is probably a bad idea since it is
part of the base system and freebsd-update(8) may want to update it.

According to pkg.conf(5):

     The repository tag	myrepo is an arbitrary string.	Reusing	the repository
     tag will cause those items	defined	in configuration files later on	the
     REPOS_DIR search path to overwrite	the equivalent settings	for the	same
     tag earlier on the	search path.  Hence the	very common idiom, used	to
     turn off the default FreeBSD configuration	shipped	in
     /etc/pkg/FreeBSD.conf.  Rather than editing that file directly, create
     /usr/local/etc/pkg/repos/FreeBSD.conf with	this content:

	   FreeBSD: { enabled: NO }

So, ask to do this way in the README and only purge the user defined
repositories that are not managed by Puppet, following the
recommendations of the man page.